src/main/java/com/mikaeru/financialmanager/domain/service/user/UserService.java
package com.mikaeru.financialmanager.domain.service.user;

import com.mikaeru.financialmanager.domain.model.user.User;
import com.mikaeru.financialmanager.domain.repository.user.UserRepository;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class UserService {

    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User findUser(String email) throws Exception {
        Optional<User> userOptional = userRepository.findByEmail(email);

        if (userOptional.isEmpty()) {
            // TODO Lançar uma exception do Tipo UserNotFoundException.
            throw new Exception();
        }

        return userOptional.get();
    }
}
